+++ CHANGELOG.md
@@ -9,6 +9,9 @@ PATCH
 ### Payments
 * [Added] Added API bindings for Scalapay.
 
+### PaymentSheet
+* [Fixed] PaymentSheet now keeps the Apple Pay EMV merchant capability aligned with China UnionPay in the final supported network list.
+
 ## 26.10.0 2026-09-08
 ### General
 * [Added] Added support for Arabic (Saudi Arabia).